+++ winpr/libwinpr/utils/cmdline.c
@@ -115,7 +115,7 @@ int CommandLineParseArgumentsA(int argc, LPCSTR* argv, COMMAND_LINE_ARGUMENT_A* 
 		sigil_index = 0;
 		sigil_length = 0;
 		sigil = (char*) &argv[i][sigil_index];
-		length = strlen(argv[i]);
+		length = (int) strlen(argv[i]);
 
 		if ((sigil[0] == '/') && (flags & COMMAND_LINE_SIGIL_SLASH))
 		{
@@ -198,9 +198,9 @@ int CommandLineParseArgumentsA(int argc, LPCSTR* argv, COMMAND_LINE_ARGUMENT_A* 
 			if (separator)
 			{
 				separator_length = 1;
-				separator_index = (separator - argv[i]);
+				separator_index = (int) (separator - argv[i]);
 
-				keyword_length = (separator - keyword);
+				keyword_length = (int) (separator - keyword);
 
 				value_index = separator_index + separator_length;
 				value = (char*) &argv[i][value_index];
@@ -278,7 +278,7 @@ int CommandLineParseArgumentsA(int argc, LPCSTR* argv, COMMAND_LINE_ARGUMENT_A* 
 					{
 						i++;
 						value_index = 0;
-						length = strlen(argv[i]);
+						length = (int) strlen(argv[i]);
 
 						value = (char*) &argv[i][value_index];
 						value_length = (length - value_index);
